Сфомированы автотесты
