+++ precode_test.go
@@ -1,0 +1,41 @@
+package main
+
+import (
+	"github.com/stretchr/testify/assert"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+)
+
+func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	totalCount := 4
+	req := httptest.NewRequest("GET", "/cafe?count=100&city=moscow", nil) // здесь нужно создать запрос к сервису
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+	assert.Equal(t, http.StatusOK, responseRecorder.Code)
+	assert.Len(t, strings.Split(responseRecorder.Body.String(), ","), totalCount)
+}
+
+func TestMainHandlerWhenRequestFormedCorrectly(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=2&city=moscow", nil) // здесь нужно создать запрос к сервису
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+	assert.Equal(t, http.StatusOK, responseRecorder.Code)
+	assert.NotEqual(t, responseRecorder.Body.String(), "")
+}
+
+func TestMainHandlerWhenCityIncorrect(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=2&city=Kazan", nil) // здесь нужно создать запрос к сервису
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+	//fmt.Println(responseRecorder.Body.String())
+	assert.Equal(t, responseRecorder.Body.String(), "wrong city value")
+}
